Funcionan las fotos!!

+++ app/src/main/java/com/lisis/charles/fastreport/DB_Accident.java
@@ -5,8 +5,8 @@ package com.lisis.charles.fastreport;
  */
 public class DB_Accident {
 
-    private int user_id;
-    private int vehicle_id;
+    private long user_id;
+    private long vehicle_id;
     private String date;
     private String location;
     private String email_addressee;
@@ -28,19 +28,19 @@ public class DB_Accident {
         this.email_addressee = null;
     }
 
-    public int getUser_id() {
+    public long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
 
-    public int getVehicle_id() {
+    public long getVehicle_id() {
         return vehicle_id;
     }
 
-    public void setVehicle_id(int vehicle_id) {
+    public void setVehicle_id(long vehicle_id) {
         this.vehicle_id = vehicle_id;
     }
